requirements.txt
asgiref==3.6.0
Babel==2.12.1
crispy-bootstrap5==0.7
Django==4.2
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-htmx==1.14.0
django-money==3.1.0
django-types==0.17.0
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==20.1.0
humanize==4.6.0
mccabe==0.7.0
psycopg==3.1.9
psycopg-binary==3.1.9
py-moneyed==2.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0
